requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0
Flask-WTF==1.2.1
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
